build(server test): fix not working server before test by using jest-dev-server

+++ server/prisma/seed/admin.js
@@ -6,30 +6,40 @@ import { prisma } from '../../src/generated/prisma-client';
 const { admin } = cfg;
 
 export default async () => {
-  try {
-    await prisma.deleteUser({ email: admin.email });
-  } catch (e) {
-    console.log('Nothing to reset on admin account');
+  const hash = await argon2.hash(admin.password);
+  const newAdmin = {
+    firstName: 'Mr.',
+    lastName: 'Admin',
+    email: admin.email,
+    password: hash,
+    role: 'ADMIN',
+  };
+
+  const user = await prisma.user({ email: admin.email });
+  if (user && user.role === 'ADMIN') {
+    console.log('\n>>> Seed on admin: user already exist');
+
+    return;
+  }
+
+  if (user && user.role !== 'ADMIN') {
+    try {
+      await prisma.deleteUser({ email: admin.email });
+      console.log('Removed old admin account.');
+    } catch (e) {
+      console.log('Failed to reset admin account:', admin.email);
+      process.exit(1);
+    }
   }
 
   try {
-    const hash = await argon2.hash(admin.password);
-
-    const newAdmin = {
-      firstName: 'Mr.',
-      lastName: 'Admin',
-      email: admin.email,
-      password: hash,
-      role: 'ADMIN',
-    };
-
     await prisma.createUser(newAdmin);
 
-    console.log('\n>>> Seed on admin');
-    console.log('Removed old admin account.');
+    console.log('\n>>> Sucesful Seed on admin');
     console.log('New admin created: ', newAdmin);
   } catch (e) {
-    console.error('error on inserting admin\n', e);
+    console.error('error on inserting admin\n');
+    console.log(e);
     process.exit(1);
   }
 };
